%  -*- TeX|UK -*- -*- Soft -*-
\name{getArea-methods}
\docType{methods}
\alias{getArea-methods}
\alias{getArea}
\alias{getArea,CompactStratification-method}

\title{Extract the Area of an Object}

\description{
    Methods for extracting the area of objects.
}

\section{Methods}{
\describe{

\item{object = "CompactStratification"}{returns the area of each stratum.}
}}

\seealso{
    \code{\link[=getRelativeArea-methods]{getRelativeArea}}
}

%\examples{
%showClass("SamplingPatternPurposive")
%}


\keyword{methods} 