%  -*- TeX|UK -*- -*- Soft -*-
\name{getCentroid-methods}
\docType{methods}
\alias{getCentroid-methods}
\alias{getCentroid}
\alias{getCentroid,CompactStratification-method}

\title{Extract Centroids}

\description{
Methods for extracting centroids
}

\section{Methods}{
\describe{

\item{object = "CompactStratification"}{returns the centers of gravity of each stratum.}
}}

\keyword{methods}
