%  -*- TeX|UK -*- -*- Soft -*-
\name{getNumberOfStrata-methods}
\docType{methods}
\alias{getNumberOfStrata-methods}
\alias{getNumberOfStrata}
\alias{getNumberOfStrata,CompactStratification-method}

\title{Extract the Number of Strata in an Object}

\description{
    Methods for extracting the number of strata of objects.
}

\section{Methods}{
\describe{

\item{object = "CompactStratification"}{returns the number of strata in a compact stratification. }
}}

\keyword{methods}
