%  -*- TeX|UK -*- -*- Soft -*-
\name{getObjectiveFunctionValue-methods}
\docType{methods}
\alias{getObjectiveFunctionValue}
\alias{getObjectiveFunctionValue-methods}
\alias{getObjectiveFunctionValue,CompactStratification-method}
\title{Extract the Objective Function Value of an Object}
\description{
    Methods for extracting the objective function value
}
\section{Methods}{
\describe{

\item{object = "CompactStratification"}{extracts the mean squared shortest distance.}
}}
\keyword{methods}
