%  -*- TeX|UK -*- -*- Soft -*-
\name{getRelativeArea-methods}
\docType{methods}
\alias{getRelativeArea-methods}
\alias{getRelativeArea}
\alias{getRelativeArea,CompactStratification-method}
\title{Extract the Relative Area of an Object}

\description{
    Methods for extracting relative areas of objects. The total area equals unity.
}

\section{Methods}{
\describe{

\item{object = "CompactStratification"}{returns the relative area of each stratum. The sum of the relative areas equals 1.}
}}

\seealso{
    \code{\link[=getArea-methods]{getArea}}
}

\keyword{methods}
