%  -*- TeX|UK -*- -*- Soft -*-
\name{getSampleSize-methods}
\docType{methods}
\alias{getSampleSize-methods}
\alias{getSampleSize}
\alias{getSampleSize,SamplingPattern-method}
\alias{getSampleSize,SamplingPatternRandomComposite-method}

\title{Extract the sample size of an object}

\description{
    Methods for extracting the sample size.
}

\section{Methods}{
\describe{

\item{object = "SamplingPattern"}{returns the sample size.}

\item{object = "SamplingPatternRandomComposite"}{returns the number of composites}
}}

\keyword{methods}
