%  -*- TeX|UK -*- -*- Soft -*-
\name{SamplingPatternPurposive-class}
\docType{class}
\alias{SamplingPatternPurposive-class}

\title{Class "SamplingPatternPurposive"}

\description{An ancestor class for storing purposive sampling patterns.}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SamplingPatternPurposive", ...)}.
}

\section{Slots}{
    \describe{
        \item{\code{sample}:}{Object of class \code{"\link[sp:SpatialPoints]{sp::SpatialPoints}"}, containing the sampling locations}
    }
}

\section{Extends}{
Class \code{"\linkS4class{SamplingPattern}"}, directly.
}

\section{Methods}{
No methods defined with class "SamplingPatternPurposive" in the signature.
}

\author{Dennis J. J. Walvoort \email{dennis.walvoort@wur.nl}, D.J. Brus, J.J. de Gruijter}

\keyword{classes}
