%  -*- TeX|UK -*- -*- Soft -*-
\name{SamplingPatternRandom-class}
\docType{class}
\alias{SamplingPatternRandom-class}

\title{Class "SamplingPatternRandom"}

\description{An ancestor class for storing random sampling patterns.}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SamplingPatternRandom", ...)}.
}

\section{Slots}{
	 \describe{
    \item{\code{sample}:}{Object of class \code{"\link[sp]{SpatialPoints}"}, containing the sampling locations}
  }
}

\section{Extends}{
Class \code{"\linkS4class{SamplingPattern}"}, directly.
}

\section{Methods}{
No methods defined with class "SamplingPatternRandom" in the signature.
}

\author{Dennis J. J. Walvoort \email{dennis.walvoort@wur.nl}, D.J. Brus, J.J. de Gruijter}

\keyword{classes}
