%  -*- TeX|UK -*- -*- Soft -*-
\name{Statistic-class}
\docType{class}
\alias{Statistic-class}
\alias{show,Statistic-method}

\title{Class "Statistic"}

\description{A superclass (ancestor class) for statistics to estimate.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Slots}{
	 \describe{
    \item{\code{description}:}{A description op the statistic}
  }
}

\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "Statistic")}: prints the statistic}
	 }
}

\author{Dennis J. J. Walvoort \email{dennis.walvoort@wur.nl}, D.J. Brus, J.J. de Gruijter}

\keyword{classes}
