%  -*- TeX|UK -*- -*- Soft -*-
\name{Stratification-class}
\docType{class}
\alias{Stratification-class}
\alias{show,Stratification-method}

\title{Class "Stratification"}

\description{Virtual class to store a spatial stratification.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "Stratification")}: a method for printing objects of class \code{Stratification}}
	 }
}

\author{Dennis J. J. Walvoort \email{dennis.walvoort@wur.nl}, D.J. Brus, J.J. de Gruijter}

\examples{
showClass("Stratification")
}

\keyword{classes}
